requirements.txt
streamlit
yfinance
duckduckgo-search
openai
numpy
pandas
plotly
seaborn
python-dotenv
